+++ packages/data-broker/src/services/dhis/getDhisApiInstance.js
@@ -5,7 +5,7 @@
 
 import { DhisApi } from '@tupaia/dhis-api';
 import { getDhisConfig, createClassExtendingProxy } from '@tupaia/utils';
-import { DhisInputSchemeResolvingApiProxy } from './translators/DhisInputSchemeResolvingApiProxy';
+import { DhisInputSchemeResolvingApiProxy } from './translators';
 
 const instances = {};
