+++ packages/data-broker/src/services/dhis/translators/index.js
@@ -1,0 +1,7 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ */
+
+export { DhisInputSchemeResolvingApiProxy } from './DhisInputSchemeResolvingApiProxy';
+export { DhisTranslator } from './DhisTranslator';
